separate connectivity and reachability code from process code
