+++ src/electron/process_manager.ts
@@ -13,18 +13,12 @@
 // limitations under the License.
 
 import {ChildProcess, execSync, spawn} from 'child_process';
-import * as dgram from 'dgram';
-import * as dns from 'dns';
-import {app} from 'electron';
-import * as net from 'net';
 import * as os from 'os';
-import * as path from 'path';
-import * as socks from 'socks';
 
-import * as util from '../www/app/util';
 import * as errors from '../www/model/errors';
 
 import {SerializableConnection} from './connection_store';
+import * as connectivity from './connectivity';
 import * as routing from './routing_service';
 import {pathToEmbeddedBinary} from './util';
 
@@ -47,8 +41,8 @@ const routingService = new routing.RoutingService();
 //  - badvpn-tun2socks.exe connects the SOCKS5 proxy to a TAP-like network interface
 //  - OutlineService configures the system to route via a TAP-like network device, must be installed
 
-let ssLocal: ChildProcess | undefined;
-let tun2socks: ChildProcess | undefined;
+let ssLocal: ChildProcess|undefined;
+let tun2socks: ChildProcess|undefined;
 let activeConnection: SerializableConnection;
 
 const PROXY_IP = '127.0.0.1';
@@ -61,14 +55,8 @@ const TUN2SOCKS_VIRTUAL_ROUTER_IP = '10.0.85.1';
 const TUN2SOCKS_TAP_DEVICE_NETWORK = '10.0.85.0';
 const TUN2SOCKS_VIRTUAL_ROUTER_NETMASK = '255.255.255.0';
 
-const CREDENTIALS_TEST_DOMAINS = ['example.com', 'ietf.org', 'wikipedia.org'];
 const SS_LOCAL_TIMEOUT_SECS =
-  2 ^ 31 - 1;  // 32-bit INT_MAX; using Number.MAX_SAFE_VALUE may overflow
-const REACHABILITY_TEST_TIMEOUT_MS = 10000;
-const DNS_LOOKUP_TIMEOUT_MS = 10000;
-const UDP_FORWARDING_TEST_TIMEOUT_MS = 5000;
-const UDP_FORWARDING_TEST_RETRY_INTERVAL_MS = 1000;
-
+    2 ^ 31 - 1;  // 32-bit INT_MAX; using Number.MAX_SAFE_VALUE may overflow
 // This function is roughly the Electron counterpart of Android's VpnTunnelService.startShadowsocks.
 //
 // Fulfills iff:
@@ -113,13 +101,13 @@ export function startVpn(
         }
         // Only perform the connectivity checks when we're not automatically connecting on boot,
         // since we may not have network connectivity.
-        return checkConnectivity(config).then((ip) => {
+        return connectivity.checkConnectivity(config, PROXY_IP, SS_LOCAL_PORT).then((ip) => {
           // Cache the resolved IP so it can be stored for auto connect.
           config.host = ip;
         });
       })
       .then(() => {
-        return checkUdpForwardingEnabled()
+        return connectivity.checkUdpForwardingEnabled(PROXY_IP, SS_LOCAL_PORT)
             .then((isRemoteUdpForwardingEnabled) => {
               isUdpSupported = isRemoteUdpForwardingEnabled;
               console.log('UDP forwarding', isUdpSupported ? 'enabled' : 'disabled');
@@ -137,8 +125,7 @@ export function startVpn(
       .then(delay(isLinux ? WAIT_FOR_PROCESS_TO_START_MS : 0))
       .then(() => {
         return routingService.configureRouting(
-            TUN2SOCKS_VIRTUAL_ROUTER_IP, config.host || '', connectionStatusChanged,
-            isAutoConnect);
+            TUN2SOCKS_VIRTUAL_ROUTER_IP, config.host || '', connectionStatusChanged, isAutoConnect);
       })
       .then(() => {
         activeConnection = {id: '', config, isUdpSupported};
@@ -187,68 +174,8 @@ function testTapDevice() {
   }
 }
 
-// Fulfills iff:
-//  - we can connect to the Shadowsocks server port
-//  - we can speak with a semi-random test site via the proxy
-function checkConnectivity(config: cordova.plugins.outline.ServerConfig): Promise<string> {
-  return lookupIp(config.host || '').then((ip: string) => {
-    return isServerReachableByIp(ip, config.port || 0)
-      .then(() => {
-        return validateServerCredentials();
-      })
-      .then(() => {
-        return ip;
-      });
-  });
-}
-// Uses the OS' built-in functions, i.e. /etc/hosts, et al.:
-// https://nodejs.org/dist/latest-v10.x/docs/api/dns.html#dns_dns
-//
-// Effectively a no-op if hostname is already an IP.
-function lookupIp(hostname: string): Promise<string> {
-  return util.timeoutPromise(
-    new Promise<string>((fulfill, reject) => {
-      dns.lookup(hostname, 4, (e, address) => {
-        if (e) {
-          return reject(new errors.ServerUnreachable('could not resolve proxy server hostname'));
-        }
-        fulfill(address);
-      });
-    }),
-    DNS_LOOKUP_TIMEOUT_MS, 'DNS lookup');
-}
-
-// Resolves with true iff a TCP connection can be established with the Shadowsocks server.
-//
-// This has the same function as ShadowsocksConnectivity.isServerReachable in
-// cordova-plugin-outline.
-export function isServerReachable(config: cordova.plugins.outline.ServerConfig) {
-  return lookupIp(config.host || '').then((ip) => {
-    return isServerReachableByIp(ip, config.port || 0);
-  });
-}
-
-// As #isServerReachable but does not perform a DNS lookup.
-export function isServerReachableByIp(serverIp: string, serverPort: number) {
-  return util.timeoutPromise(
-    new Promise<void>((fulfill, reject) => {
-      const socket = new net.Socket();
-      socket
-        .connect(
-          { host: serverIp, port: serverPort },
-          () => {
-            socket.end();
-            fulfill();
-          })
-        .on('error', () => {
-          reject(new errors.ServerUnreachable());
-        });
-    }),
-    REACHABILITY_TEST_TIMEOUT_MS, 'Reachability check');
-}
-
 function startLocalShadowsocksProxy(
-  serverConfig: cordova.plugins.outline.ServerConfig, onDisconnected: () => void) {
+    serverConfig: cordova.plugins.outline.ServerConfig, onDisconnected: () => void) {
   return new Promise((resolve, reject) => {
     // ss-local -s x.x.x.x -p 65336 -k mypassword -m aes-128-cfb -l 1081 -u
     const ssLocalArgs = ['-l', SS_LOCAL_PORT.toString()];
@@ -302,112 +229,10 @@ function startLocalShadowsocksProxy(
   });
 }
 
-// Resolves with true iff a response can be received from a semi-randomly-chosen website through the
-// Shadowsocks proxy.
-//
-// This has the same function as ShadowsocksConnectivity.validateServerCredentials in
-// cordova-plugin-outline.
-function validateServerCredentials() {
-  return new Promise((fulfill, reject) => {
-    const testDomain =
-      CREDENTIALS_TEST_DOMAINS[Math.floor(Math.random() * CREDENTIALS_TEST_DOMAINS.length)];
-    socks.createConnection(
-      {
-        proxy: { ipaddress: PROXY_IP, port: SS_LOCAL_PORT, type: 5 },
-        target: { host: testDomain, port: 80 }
-      },
-      (e, socket) => {
-        if (e) {
-          reject(new errors.InvalidServerCredentials(
-            `could not connect to remote test website: ${e.message}`));
-          return;
-        }
-
-        socket.write(`HEAD / HTTP/1.1\r\nHost: ${testDomain}\r\n\r\n`);
-
-        socket.on('data', (data) => {
-          if (data.toString().startsWith('HTTP/1.1')) {
-            socket.end();
-            fulfill();
-          } else {
-            socket.end();
-            reject(new errors.InvalidServerCredentials(
-              `unexpected response from remote test website`));
-          }
-        });
-
-        socket.on('close', () => {
-          reject(new errors.InvalidServerCredentials(`could not connect to remote test website`));
-        });
-
-        // Sockets must be resumed before any data will come in, as they are paused right before
-        // this callback is fired.
-        socket.resume();
-      });
-  });
-}
-
-// Verifies that the remote server has enabled UDP forwarding by sending a DNS request through it.
-function checkUdpForwardingEnabled(): Promise<boolean> {
-  return new Promise((resolve, reject) => {
-    socks.createConnection(
-      {
-        proxy: { ipaddress: PROXY_IP, port: SS_LOCAL_PORT, type: 5, command: 'associate' },
-        target: { host: '0.0.0.0', port: 0 },  // Specify the actual target once we get a response.
-      },
-      (err, socket, info) => {
-        if (err) {
-          reject(new errors.RemoteUdpForwardingDisabled(`could not connect to local proxy`));
-          return;
-        }
-        const dnsRequest = getDnsRequest();
-        const packet = socks.createUDPFrame({ host: '1.1.1.1', port: 53 }, dnsRequest);
-        const udpSocket = dgram.createSocket('udp4');
-
-        udpSocket.on('error', (e) => {
-          reject(new errors.RemoteUdpForwardingDisabled('UDP socket failure'));
-        });
-
-        udpSocket.on('message', (msg, info) => {
-          stopUdp();
-          resolve(true);
-        });
-
-        // Retry sending the query every second.
-        // TODO: logging here is a bit verbose
-        const intervalId = setInterval(() => {
-          try {
-            udpSocket.send(packet, info.port, info.host, (err) => {
-              if (err) {
-                console.error(`Failed to send data through UDP: ${err}`);
-              }
-            });
-          } catch (e) {
-            console.error(`Failed to send data through UDP ${e}`);
-          }
-        }, UDP_FORWARDING_TEST_RETRY_INTERVAL_MS);
-
-        const stopUdp = () => {
-          try {
-            clearInterval(intervalId);
-            udpSocket.close();
-          } catch (e) {
-            // Ignore; there may be multiple calls to this function.
-          }
-        };
-
-        // Give up after the timeout elapses.
-        setTimeout(() => {
-          stopUdp();
-          resolve(false);
-        }, UDP_FORWARDING_TEST_TIMEOUT_MS);
-      });
-  });
-}
-
 // Checks whether UDP forwarding support has changed and restarts tun2socks if it has.
 async function handleUdpSupportChange(onDisconnected: () => void): Promise<void> {
-  const isUdpSupported = await checkUdpForwardingEnabled().catch(e => false);
+  const isUdpSupported =
+      await connectivity.checkUdpForwardingEnabled(PROXY_IP, SS_LOCAL_PORT).catch(e => false);
   if (isUdpSupported !== activeConnection.isUdpSupported) {
     console.info(`UDP support changed (${activeConnection.isUdpSupported} > ${isUdpSupported})`);
     activeConnection.isUdpSupported = isUdpSupported;
@@ -415,23 +240,6 @@ async function handleUdpSupportChange(onDisconnected: () => void): Promise<void>
     const tunDeviceName = await getTunDeviceName();
     await startTun2socks(tunDeviceName, activeConnection.isUdpSupported, onDisconnected);
   }
-}
-
-// Returns a buffer containing a DNS request to google.com.
-function getDnsRequest() {
-  return Buffer.from([
-    0, 0,                             // [0-1]   query ID
-    1, 0,                             // [2-3]   flags; byte[2] = 1 for recursion desired (RD).
-    0, 1,                             // [4-5]   QDCOUNT (number of queries)
-    0, 0,                             // [6-7]   ANCOUNT (number of answers)
-    0, 0,                             // [8-9]   NSCOUNT (number of name server records)
-    0, 0,                             // [10-11] ARCOUNT (number of additional records)
-    6, 103, 111, 111, 103, 108, 101,  // google
-    3, 99, 111, 109,                  // com
-    0,                                // null terminator of FQDN (root TLD)
-    0, 1,                             // QTYPE, set to A
-    0, 1                              // QCLASS, set to 1 = IN (Internet)
-  ]);
 }
 
 function startTun2socks(
